+++ TotalImage.IO/FileSystems/FileSystem.cs
@@ -85,6 +85,6 @@ namespace TotalImage.FileSystems
         /// <summary>
         /// The minimum allocatable unit size on the file system
         /// </summary>
-        public abstract uint AllocationUnitSize { get; }
+        public abstract long AllocationUnitSize { get; }
     }
 }
