+++ TotalImage.IO/FileSystems/RAW/RawFileSystem.cs
@@ -35,6 +35,6 @@ namespace TotalImage.FileSystems.RAW
         public override long TotalSize => _stream.Length;
 
         /// <inheritdoc />
-        public override uint AllocationUnitSize => 512;
+        public override long AllocationUnitSize => 512;
     }
 }
